+++ src/components/app/index.ts
@@ -19,6 +19,13 @@ import createService, {
 import { loadYaml } from "../../utils/getEnvironmentComponent";
 import { updateMetadata } from "../../utils/updateMetadata";
 import { merge } from "../../utils/merge";
+import { addPostgresUserSecret } from "../../utils/addPostgresUserSecret";
+import { addWaitForPostgres } from "../../utils/addWaitForPostgres";
+
+type AliasParams = {
+  hosts: string[];
+  destination: string;
+};
 
 export type AppConfig = DeploymentParams &
   ServiceParams &
@@ -28,6 +35,8 @@ export type AppConfig = DeploymentParams &
     domain: string;
     labels: Record<string, string>;
     ingress: boolean;
+    withPostgres: boolean;
+    withRedirections?: AliasParams;
   };
 export const create = (
   name: string,
@@ -69,6 +78,26 @@ export const create = (
     name,
   });
   manifests.push(deployment);
+
+  // add postgres secret and initContainer
+  if (envParams.withPostgres) {
+    addPostgresUserSecret(deployment);
+    addWaitForPostgres(deployment);
+  }
+
+  // add a redirection ingresses, production only
+  if (env.env === "prod" && envParams.withRedirections) {
+    const { hosts, destination } = envParams.withRedirections;
+    const redirectIngress = createIngress({
+      name: `${name}-redirects`,
+      hosts,
+      secretName: `${name}-redirects`,
+      annotations: {
+        "nginx.ingress.kubernetes.io/permanent-redirect": `https://${destination}$request_uri`,
+      },
+    });
+    manifests.push(redirectIngress);
+  }
 
   /* SEALED-SECRET */
   // try to import environment sealed-secret